Merged PR 724765: Add self-recovery capabilities for the blob L3 cache if a faulty state is reached wrt eviction
If we reach a point where a content pin was elided, but we later find that the content is not there, builds will fail with a materialization error. We can reach this faulty state because the blob storage got corrupted (e.g. somebody deleted content manually) or because of bugs in the pin eliding strategy. The main problem with this is that subsequent builds will likely fail until the associated content hash list is deemed to be evicted (which can be many days), so pins start to not be elided anymore.

The currently running build experimenting materialization failures is beyond saving. But here we add a mechanism to clear out the last content pinned time from the associated fingerprint, so subsequent builds will preventively pin again and won't fail.

Related work items: #2075982