Merged PR 732070: Fix processor usage percentage for pips in linux
The processor usage percentage for pips in Linux was miscalculated, always resulting in a value of 0. This error prevented us from utilizing the dynamic weight based on historical data.

We extract the user and kernel time from the stat file, where the unit is measured in ticks. Therefore, we need the 'ticks per second' value from the system configuration to convert the ticks to milliseconds. To obtain the precise 'ticks per second' for a particular machine, we invoke sysconf with the _SC_CLK_TCK flag. However, the value for _SC_CLK_TCK was incorrect; it was set to 1 instead of 2, leading us to receive 64k instead of the expected 100. It's worth noting that the value 1 used with the sysconf function corresponds to _SC_ARG_MAX, which queries the maximum length of the arguments for the functions

Related work items: #2086626