Merged PR 733863: Ensure the empty file doesn't make it to AzureBlobStorageContentSession
`AzureBlobStorageContentSession` inherits from `ContentSessionBase` and does not override the default bulk operations. This means that by ensuring `ContentSessionBase` does not ever put or pin the empty file, we ensure the same for the blob content session.

There are some specific optimizations for the empty file in the content session implementation (such as `TouchAsync`), but I won't remove them until we have made sure that they aren't being triggered ever.

Finally, found some flakyness in our tests due to the chosen working directory and concurrent tests. Started using `DisposableDirectory` to ensure uniqueness + cleanup. This is fine now that this does not run on dev machines and checkpointing allows for post-mortem/debugging.