Merged PR 735180: Added additional check and logging to WorkspaceDownloadResolver to gather more information about the Linux execute permission bug.
The executable file for node in linux sometimes when retrieved from cache is not having the required execute bit set, causing a couple of builds to fail.

This is not a complete fix , as the hack provided is just for one of the assumptions that the bit might not have been set right when the package has been downloaded and extracted to the specific directory.

Added logging and check to ensure that this is identified when the package has been downloaded and extracted. This involves the use of DownloadWorkspaceResolver.cs, which in turn uses Extractor to extract the packages.

Tested for these unit test cases:
1) When the file is not existent
2) When the wrong file path is given
3) When the bit set is right or not

Related work items: #2073919