Merged PR 735405: Add an option to honor directory casing for dynamic outputs
We hit an incident where JS builds in Windows are actually case sensitive. On cache replay, we only make sure the file name preserves casing, but not the directories. In this particular case, absent probes happened before outputs were produced where directories in the probe contained the wrong casing. This polluted the path table and made BuildXL replay the files with the wrong directory casing.

This change extends the filename case preserving logic to the relative path for dynamic outputs. The change is happening under a flag for deployment reasons, but ideally, we can validate this is not breaking and make this the default/remove the flag.

The change is essentially about making sure the right casing goes into the cache metadata for dynamic output paths, and on cache replay, preserving the casing we found until we reach the cache layer.

Related work items: #2081323, #2083467