Merged PR 735862: Fix InputTrackerAnalyzer - Deserialization bug in ReadOnlyFixedBytes.ReadFrom()
The InputTrackerAnalyzer was crashing because its deserialization of PreviousInputs was getting out of sync from the file.

The underlying problem was a bug in ReadOnlyFixedBytes.ReadFrom(). When reading from a stream, it was not checking how many bytes got returned from the read. Depending on the implementation of the stream, additional read calls may be required to return all of the requested bytes. This is exactly what happens with DeflateStream, which is used when the graph files are compressed in CloudBuild.

Related work items: #2100196