Merged PR 735490: Classify WriteOnAbsentPathProbe as missing dependency in DFA summary
Write on absent path probe can happen in two cases - probe followed by write, and write followed by probe. Also, in distributed builds, the order of pip execution might differ from the order the pip undergo DFA checks. Currently, we are simplifying WriteOnAbsentPathProbe to "write", because of that if a probing pip loses the race for DFA check, the DFA summary states that the probing pip did a write (which is incorrect).

With this PR, we'll simplify WriteOnAbsentPathProbe to "missing dependency". This way the summary message will also align with the suggestion (add dependency) that we give in the verbose message that we log for each WriteOnAbsentPathProbe.

Related work items: #2097763