Merged PR 730467: Do not lie about cause of PlaceFailure
The error message that the top-level cache error message that appears in batmon (“No replicas currently exist in content tracker for hash”) is often a lie. It appears when there are location is LLS, we try them all and fail, and we fall back to GCS (or Redis) and find the same locations. This happens because, in an effort not to repeat our failures, we remove locations that we already tried from the fallback location list. If there were no new locations in GCS (or Redis), that list is then empty, which our error condition logic then interprets as “we don’t have any locations” and therefore writes the misleading error message.

Scenarios:

- Filtered all inactive machines:
        - **Before**: No replicas currently exist in content tracker for hash XXXX
        - **After**: Filtered all locations for XXXX because they were inactive machines

- Failed against LLS and against GCS:
        - **Before** we showed the fallback error message
        - **After** we will show the original error message

    NOTE: Maybe instead of the original error we want to keep both (concatenate them). We can discuss this but I think it's better to keep it simple.

- Removed all locations from GCS (because there were no new locations):
        - **Before**: No replicas currently exist in content tracker for hash XXXX
        - **After**: Replicas in LLS content tracker for hash XXXX but failed to copy all

    NOTE: We will not show this error because we will override it with the previous rule.