Merged PR 739191: Change Ephemeral OpenStream to avoid unnecessary disk IO
This PR changes the OpenStream implementation in EphemeralContentSession to avoid doing unnecessary IO.

We're currently always creating a temporary file and opening a stream to it. This allows us to avoid some of this by following the fallback order instead, which will allow us to avoid that extra file almost every time.

This PR also changes the FileRealizationMode for PutTrustedFile to Move when doing a copy.