Merged PR 739463: Keep traversing candidates when hitting an evicted content hash list
When hitting an absent content hash list, returned as a two-phase lookup candidate, we were just making that a cache miss without traversing other potential candidates. In CloudBuild, or more generally, when there is a cache eviction process that has semantic knowledge of how metadata elements are correlated, this case should be very rare. On the contrary, when doing naive cache eviction based on last access time (the case of the blob cache), this is not as uncommon. This PR changes the behavior in this case, to keep traversing candidates.

Observe this shouldn't have any visible impact on CB customers (because this case is very rare there), but it should add more robustness to the blob/ADO case.

Related work items: #2100307