Merged PR 739610: Update SBOM pipeline command line args and dotnet version
This pipeline runs the update tool in Domino.ReleaseManagement to update SBOM packages. It will search the a feed in 1essharedassets (rather than the default feed we use in cloudbuild), because it is the upstream for all SBOM packages. The ADO API does not show versions not in your feed, so it's only possible to get the latest version by looking at all the versions in the upstream feed.

Related work items: #2100959