Merged PR 737831: Collect GIT remote repo info for dev builds
Created a class GitInfoManager as there might be a possibility that another implementation is also used to collect Git remote repo URL.
This class aims to collect the remote repo URL from a dev build.
    /// To gather necessary details from a Git remote repository, two methods were discussed:
    /// Parsing of the git/config file or employing git.exe.
    /// Initially, we will leverage the git/config file for this purpose. However, if this method doesn't yield the anticipated results, we will resort to the second approach using git.exe as a backup plan.

Created a set of unit tests in GitInfoManager ests to check for the functionality.
Added a sample git config file in the comments of the class to help visualize the contents of the file better.
Added a counter to record the time to perform the entire operation.

Related work items: #2095396