Merged PR 739783: [Blob cache] Use DownloadStreamingAsync instead of DownloadContentAsync for TouchAsync
Replace DownloadContentAsync  with DownloadStreamingAsync in the implementation of TouchAsync. Considering this blog post https://github.com/Azure/azure-sdk-for-net/issues/22022 and browsing the implementation of both operations, DownloadStreamingAsync is slightly more lightweight. Disclaimer: network operations are handled in the same way, DownloadStreamingAsync just avoids some stream copying which DownloadContentAsync performs - and content is something we don't actually interact with when doing a touch operation.