Merged PR 740667: Chunker should share pool across instances
Chunker should share pool across instances

When building with memobuild, msbuild ballooned to 24GB of RAM and that caused the compiler to run out of memory.  When I looked at the dump, 19GB of it was "dead" and was from DeterministicChunker's pools that were not used. Sergey confirmed that the right way to use pool is via a static.