Merged PR 742394: Revert 'Mark additional QTest Arguments file as untracked'
Turns out that if this file is untracked, QTest pips are not invalidated with changes to this file, causing QTest pips to not run.

This is a feature request from @<Ibrahim Abdullah> . When passing additional args file to QTest, Office is trying to use a common file located centrally. This is causing DFA when creating the QTest process.
Untracking the read-only file will resolve this issue.

Reverts !740843