Merged PR 742087: Remove version restriction from blob options and centralize HttpClient
Centralizes the storage options further to make sure that all across cache we're not going over the connection limit. Also removes the version limit we're placing on the storage sdk because of our old version of azurite.

This also fixes a bug when using managed identities that would cause the user to not utilize the singleton HttpClient, potentially running into the FD limit