Merged PR 742819: Expand frontiers of runtime cache miss analysis
Runtime cache miss analysis, by default, is only performed on frontier pips. I.e., if an analysis is performed on a pip, then no analysis is performed to its downstream pips. This is done to reduce the number of analysis as it can be fairly expensive.

We call the pips where analysis is performed frontier pips.

The issue is if when the frontier pip is an uncacheable pip, then the cache miss analysis will tell that the miss is due to the pip is configured uncacheable, but his information is not so useful because the developers already know that this pip will always have a cache miss. Developers are interested on cache misses that occurred on the downstream pips instead.

To this end, when an analysis is done on an uncacheable pip, we don't mark its downstream pips as visited. In such a way, the analysis can be done on the downstream pips. In essence, we simply expand the notion of frontier pips when there is uncachable parent.

Related work items: #2109399