Merged PR 743336: Crash fix: Avoid checking frontend preconditions when cancellation requested
All instances of the crash in the linked bug were proceeded by a cancellation. The preconditions about the state of config file processing are not valid during cancellation as various tasks bail out at different times. This change short circuits the crashing codepath earlier so it won't make it to the contract violation

Related work items: #2110755