Merged PR 744905: Render namespaces and identifiers taking into consideration escaping rules
The Dscript parser interns double underscore identifiers by adding an additional underscore. This PR makes sure that when identifiers are rendered, they are 'unescaped' to remove the additional underscore.

The change is mostly focused on namespaces and variable names, since those are the most prominent in log files and they become part of the pip provenance. Other cases are pending, but they are more unlikely to show up in the logs.

The PR also cleans up some leftover code related to escaping identifiers (the feature flag driving this behavior has been turned into a no-op a long time ago, but we still had some traces of the opt-in behavior)

Related work items: #2088726