Merged PR 745376: [Linux sandbox] Avoid interference between primary and secondary access report consuming threads
Pull request (!715221) introduced a mechanism where we read from a secondary FIFO that is meant to be a "fast lane" for high-priority messages to reach the engine faster. This introduced a second message-consuming thread that contributed to the same data structures as the primary one (notably, the  `m_accessReportProcessingBlock` ActionBlock).

When we observe that all the processes have exited, we send a sentinel value through the FIFO(s) that signal to these consuming threads that the pipe will not receive additional messages: this is sound, because the only writers to the pipes are the monitored processes that we have just seen exiting, so it is a fact that no new reports will come through them (assuming of course that we have kept good track of the full process tree).

At the moment the sentinel is pushed to the FIFO, the FIFO itself might still hold messages that have not been read, because the consuming threads are still reading from the pipe and posting the messages to the processing block one by one: the sentinel will be read after all the pending data in the pipe.

The problem is that both the primary and secondary threads are using the same action block to post the received messages: this means that the first consuming thread that encounters `EndOfReports` (typically the "secondary" thread, which doesn't have any pending messages and just instantly receives the sentinel) will mark the "shared" action block as complete. But the primary thread is still processing pending messages because the FIFO has a lot of data!

All of those messages that are consumed by the primary thread after the secondary thread has seen the sentinel and marked the action block as complete never make it to processing, because the action block is now complete and not accepting messages: crucially, the `Post` method in `ActionBlockSlim` silently drops the message when not explicitly passing `throwOnFullOrComplete`, so we were not noticing that this was happening.

This PR introduces a new action block for the secondary thread, abstracting the "reporting channels" into a new object, and makes sure we await both primary and secondary completions before marking the reporting as "done".

Related work items: #2101335