Merged PR 744599: Ensure execute permissions bit is set for destination file for copy file pip scenarions
This PR covers two use cases.
Case 1 : The content source is already in the cache, so the execution of the copy file is simply materializing the content from the cache to the specified target.

Case2 : When the source of the copyfile pip is a source file.

Case3: When the source of the copy file pip is a symlink.

Previous draft PR
https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal/pullrequest/744302?path=/Public/Src/Utilities/Storage/FileMaterializationInfo.cs