Merged PR 745336: [Linux sandbox] Make report processing more robust
The active process checker runs on a separate thread and can detect 0 running orphan processes (with respect to the active process table we keep in memory). However, there is a race between an active process we are aware of launching a new process and then finishing right away, and our ability to process the just sent 'process start' report. There could be actually a window where the active process checker will reach a 0 count, but there is a process start report being processed, at the point where the active process table is not updated yet.

When the active process checker reaches 0, it sends the end of report sentinel. This implies that as soon as we see the sentinel on the other side of the FIFO, we'll stop processing messages. This can make us miss messages if a start process report arrived before the sentinel, but after the active process checker saw 0.

This PR delays the decision to stop processing messages to the point where the sentinel is processed and we still see 0 active processes. Otherwise, we just ignore the sentinel and wait for the active process checker to send it again once we reach 0.