Merged PR 747107: Cap the proc usage of processes to 1000 in linux
cpuUsageInPercent sometimes becomes way higher than the number of physical threads of the machine.  There is an issue about getting user and kernel time for processes in some cases. It is under investigation. For now, we cap the cpuUsageInPercent to 1000 (weight:10).

I also got rid of the asymmetric merge for perf historic data.

Related work items: #2116515