Merged PR 747970: Interpose vfork
We were missing some accesses because of not interposing vfork. vfork is designed to only to exec after it, but the interposed version of exec might need to resolve some readlinks that we were not seeing before the corresponding process start, so those never had their corresponding 'start process' associated.

In addition to that, doing additional things on exec is theoretically incorrect under vfork, where only exec* and _exit are allowed.

This PR interposes vfork by calling regular fork. vfork is obsolete at this point, and modern Linux distributions should have none or negligible perf differences. And by calling vfork we also close the theoretical gap of resolving symlinks during exec interposing.

As part of this PR, refactored some of the linux compiler SDK out of the sandbox into a common place (which is now is also used for tests)