Merged PR 748389: Add Result Filter to PipExecutionPerformance Analyzer
Often you only want to get stats for the PIPs that executed instead of all PIPs.
This is especially true for very large builds where dumping all PIPs can result in a multi-gigabyte JSON file.

This change adds the option `/resultFilter:X` or `/rf:X` option to specify which `ExecutionLevel` to emit to the JSON file.