Merged PR 749202: Create BuildXL.Cache.BlobLifetimeManager.Library NuGet package
Create BlobLifetimeManager NuGet package. I've decided that, given that we'll need to get secrets and logging from CB services, it's easier to reimplement the executable in the CB repo. The executable is a very shallow shell around the libraries, so by exporting the libraries as a NuGet package, it should be very easy to do.