Merged PR 749197: Do not scrub source sealed directories
The scrubber examines all scrubbable mounts and delete all artifacts that are not a part of a build. Such mounts don't have restriction on what kind of sealed directories user can define there. Because of that, a user can define a source sealed directory (or a top-level SSD) under a scrubbable mount, and the scrubber will happily process that directory and clean it from all the source files under it at the beginning of a build.

This PR makes scrubber ignore paths that are under define source sealed and top-level source sealed directories.

Related work items: #2110268