Merged PR 748148: Use pip groups to enforce pip specific properties on selected pips.
Added a class to handle this functionality.
Added a property called /pipProperty to capture all the properties specific to a pip.
Added a unit test to check this change, covers an edge case of spaces as well.

Created PipPropertyAndValue class which has propertyName, PipSemistableHash and PropertyValue as its proeprties.
After we parse the arguments we parse the pipPRoperty argument to compute a readonly list of PipPropertyAndValue instances.
This collection is used to map pipProperties with their respective ids and properties when the engine is initialized and this map is accessed via public method for accessing various properties and their ids.