Merged PR 750373: [Blob L3 GC] Delete untracked containers after they are no longer used
Delete untracked containers after they are no longer used. The reason this is important is that we don't want to completely ignore this, since otherwise storage accounts will potentially grow unbounded.

The way we're achieving this is by tracking the last access time of all namespaces while we are reading the change feed. After we are done processing the change feed, we then check the differences between what's configured and which containers actually exist, and check if the container has been unused for long enough to be deleted.

It's important to not just delete things without checking the last access time since otherwise we might break an ongoing experiment in custom namespaces.