Merged PR 749965: Don't delete reparse point every build which invalidates incremental scheduling
BuildXL creates a stable "BuildEngineDirectory" as a reparse point inside the EngineCache pointing to where the real, potentially unstable, build engine directory is. This is referenced by various tools bundled with BuildXL since it is a stable path build over build.

The functionality uses `FileUtilites.TryCreateReparsePointIfTargetsDoNotMatch()` to create the reparse point. Previously, that function would delete and recreate the reparse point ever time, even if it was just creating the same thing again. That throws off incremental scheduling since the reparse point is getting deleted. This change makes that function actually noop when no change is needed.