Merged PR 750514: Add timeout for PipQueue draining on cancellation
Some internal errors happen when interactions with external systems are badly degraded. One example is that cache operations may be taking extremely long to complete. Since cancellation tokens do not transit through the entire cache layer, the above requested cancellation may not actually take effect for upwards of hours. In these cases, the build essentially still hangs until externally cancelled.

To avoid this, add a timeout for the amount of time the PipQueue draining thread will wait for the completion of pending tasks upon cancellation. This might leave the application in a weird state, but it gives a better shot of shutting down and flushing telemetry in a timely manner.

Related work items: #2108369