Merged PR 753272: Catch and log all exceptions in BxlAnalyzer
Currently we only catch / log only InvalidArgumentException that we throw when user provided incorrect arguments, and we rely on individual analyzers to catch any other exceptions that occur during their invocation. We've observed cases where the unobserved exceptions were not properly handled by the calller of an analyzer. This PR expands the outer try/catch, so we always process / log the exception (mainly to facilitate future debugging).

Related work items: #2116741