Merged PR 753708: Fix manifest event processing
The manifest events are sent along with the pip results. However, there is a cap for the number of pip results to be sent. That's why, some of the ipc pip results might be sent later without the manifest events. Those ipc pip results (which are sent later) can be processed earlier than the ipc results that are sent along with the manifest events because the pip results sent with the manifest events are waiting for the manifest events to be processed.

To fix this issue, the orchestrator does not receive any new messages from the same worker before we process the manifest events that have been sent.

Related work items: #2043642