Merged PR 729327: Wait for pip steps that might not have been queued before signaling PipQueue completion
If the orchestrator calls Exit on a worker after fire-forgetting all the materializations, we might still be processing materialization requests and in a stage where they haven't reached the pip queue. We should avoid marking the pip queue as complete if this is the case, so let's wait for all the pip requests to be scheduled (namely, inlined or queued into the pip queue) before doing that.