Merged PR 754347: Unify handling of default configurations
This PR creates a new entry point for getting a mutable config object. The returned config object has its default values set based on the current infra. Our current defaults (values set in constructors) is 'dev' configuration. Config for Ado and CloudBuild are set on top of it.

So, the flow of a config object looks like this:

Start with infra-specific defaults
 -> override with command line args
 -> append config from specs ('append' is the existing behavior)
 -> validate
 -> apply infra-specific overrides.

The PR also has some refactoring of the `if (InCloudBuild())` checks. A portion of them were turned into a dedicated config options and assigned appropriate default values. The rest, unfortunately, is either challenging to refactor (usually due to visibility / accessibility), or the impact of said refactoring is not clear.

Related work items: #2123456