Merged PR 748857: Add a cache miss analysis mode that uses latest git commit hashes as fingerprint store key candidates
BuildXL looks up cache miss data based on a priority list of keys configured by the user. In practice, the best general setup is to use a chain of git parent commits. This gets replicated in logic that invokes bxl.exe today for some of our customers. This PR provides this as a default experience without the end user needing to manually perform it.

The approach is heuristic, trying to pick candidate keys that might have been used to publish a fingerprint store for a build as "close" as the one running.

For this, the candidates are picked by

1. Get latest 5 commits from the current HEAD. We should get a match if this branch was built before, for example for a PR we expect one build per 'iteration'. Because an iteration can have multiple commits, we should not use a very small number (like, say, this commit and the one before), but let's not go overboard either (hence 5).
2. For any additional additional branches (indicated by the user in the argument), the assumption is that they run build against every commit (this should typically be the `main` branch, or the target branch for a PR).
  a) We retrieve the merge-base from HEAD and that branch. We use as keys the hashes for the 3 commits immediately previous to the merge-base.
  b) Finally, as a last resort, we use the latest 3 hashes from that branch as a last resort, assuming again that it was built recently

An example: for a PR (that has branched from `h5` and committed a number of times) merging against main, with this topology:

```
                                 .--b5--b4--b3--b2--b1-- b0    [PR branch]
                               /
             h8 -- h7 --h6--- h5 -- h4 -- h3 -- h2 -- h1 -- h0     [main]
```

The hashes for the keys would be (as described above)
1)`b0, b1, b2, b3, b4, b5`
2a) `h5, h6, h7`
2b) `h0, h1, h2`

Duplicates are removed, so something like this:

```
                                 .--b1-- b0    [PR branch]
                               /
             h8 -- h7 --h6--- h5 -- h1 -- h0     [main]
```

would result in the cadidates being `b0, b1, h5, h6, h7, h0, h1`.

Related work items: #2116194