Merged PR 757351: Properly dispose pip artifact state
This PR ensures that the pooled pip artifact state is disposed properly by setting all bool variables to their default state (in this case, all of them are false).

The improper disposal has *non-deterministically* caused build with incremental scheduling to fail with
```
ContractException: Assumption failed (false): Attempted to query the content hash for an artifact which has not passed through SetFileArtifactContentHash
```

When the scheduler determines that a partial/full sealed directory should be incrementally skipped, the scheduler calls `TryRegisterOutputDirectoriesAndHashSharedOpaqueOutputsAsync` of `FileContentManager`. This method tells the `FileContentManager` to marks the sealed directory as materialized. However, the marking is conditioned on `PipArtifactsState.EnforceOutputMaterializationExclusionRootsForDirectoryArtifacts`. Because `PipArtifactsState` is a pooled object, then there's a chance that we used an existing object from the pool.

The bug is `PipArtifactsState.EnforceOutputMaterializationExclusionRootsForDirectoryArtifacts` is not set to its default value when it is returned to the pool. Thus, when it is reused during the call to `TryRegisterOutputDirectoriesAndHashSharedOpaqueOutputsAsync`, the value of `PipArtifactsState.EnforceOutputMaterializationExclusionRootsForDirectoryArtifacts`  can be `true`, which makes the sealed directory not marked as materialized.

If the sealed directory is not marked as materialized, then during the input materialization step, the scheduler will try to materialize the sealed directory. For this materialization, all non-source files, including the ones in the sealed directory, must have their hashes known by the `FileContentManager`. Otherwise, we will get the above `ContractException`.

Related work items: #2122493