Merged PR 760658: Added retry for download pips
Nuget downloader uses Azure Credential Provider to access the credentials required for downloading the nuget packages.
However it does in a generalized way and does not distinguish between the various network and authentication issues.
Added handling for this and a non-zero exit code of 2 is returned to handle the retry of pips.