Merged PR 761308: Address Detours message mismatched count - use another semaphore to count successfully sent Detours messages
In short, the issue is caused by tools, e.g., node.exe/npm, that seem to have non-deterministic file access behavior, in the sense that the process termination can happen in the middle of opening a handle, in particular when Detours wants to send reported file access.

Since it's hard to handle this behavior without transacted/atomic IPC, we address this issue by "report any mismatch, but only make it as an error if Detours says that it has successfully sent a message, but the sandbox never received it"

See #2131626 for details.

Thanks to [BenWit](mailto:benwit@microsoft.com) for the deep investigation and analysis on the file accesses level that shows the "near-to-root" cause of this issue, and allows us to work on a remedy as in this PR. See [Bug 2135398](https://dev.azure.com/mseng/1ES/_workitems/edit/2135398) for details of Ben's great work.

TL;DR

| Abbreviation  | Description |
| ------------- | ------------- |
| FA  | file access   |
| FAR | file access report |
| PreSem | Semaphore used to keep track of sent/received message. Incremented by SendReport, and decremented by Receiver |
| PostSem |  Semaphore used to count the number of successfully sent message. Incremented by SendReport, and none decrements it |

**Old behavior**
```cpp
void SendReport(FileAccessReport fileAccessReport) {
 A:
 B: ReleaseSemaphore(PreSem, 1, NULL);
 C:
 D: WriteFile(pipeHandle, fileAccessReport);
 E:
}
```

Termination that happens at any of these points A, B, C, D, E
- indicates that the tool may have a nondeterministic file-access behavior, and
- indicates that the file access is potentially unimportant/irrelevant.

A:
- There's FA, no sem release, no FAR sent
- BXL is happy, i.e., it doesn't complain about mismatched FAR count because no sem has been released.
- Hard to detect the termination.

C:
- There's FA, PreSem is released, no FAR sent.
- BXL will complain about mismatched FAR count because PreSem has been released, but
   is not waited on the managed sandbox due to no FAR received (no FAR was sent).
- Can be detected if PreSem.Count > 0

D:
- as in C

E:
- There is FA, PreSem is released, FAR is sent.
- Managed sandbox receives it.
- BXL is happy, because PreSem is waited on when the managed sandbox receives it.

**New behavior**

```cpp
void SendReport(FileAccessReport fileAccessReport) {
 A:
 B: ReleaseSemaphore(PreSem, 1, NULL);
 C:
 D: bool success = WriteFile(pipeHandle, fileAccessReport);
 E:
 F: if (success) { ReleaseSemaphore(PostSem, 1, NULL); }
}
```

PostSem is used to indicate that a FAR is sent successfully.

A:
- As old behavior

C:
- As old behavior, but BXL will not err, instead it will warn about potential non-determinism.

D:
- As in C

E:
- BXL will not err, but will warn about potential non-determinism, because number of received FARs > PostSem.Count.

G:
- There's FA, PreSem is released, FAR is sent, PostSem is released.
- Managed sandbox receives it.
- BXL is happy, because PreSem is waited on when the managed sandbox receives it.

Because PostSem is used to indicate that a FAR is sent successfully, the error condition is
1. PreSem.Count > 0 // some FAR has not been received by the managed sandbox.
2. and number of received FARs < PostSem.Count //  and those FAR were successfully sent; missing FARs.